+++ cypress/e2e/form-level/form-level.cy.js
@@ -177,12 +177,12 @@ describe('Home page', () => {
       cy.contains('nav.navbar button', 'Preview').scrollIntoView().click();
       cy.contains('div[role="tab"]', 'View Rendered Form').scrollIntoView().click();
       cy.get('wc-lhc-form').should('be.visible', true);
-      cy.get('#1\\/1').as('acInput').should('have.value', 'd2 - 2');
+      cy.get('#1\\/1').as('acInput').should('have.value', 'd2');
       cy.get('@acInput').focus();
       cy.get('#completionOptionsScroller').as('acResults').should('be.visible');
       cy.get('@acResults').find('ul > li').as('acListItems').should('have.length', 2);
       cy.get('@acListItems').first().click();
-      cy.get('@acInput').should('have.value', 'd1 - 1');
+      cy.get('@acInput').should('have.value', 'd1');
       cy.contains('mat-dialog-actions > button', 'Close').click();
     });