+++ src/app/lib/lforms-widget-registry.ts
@@ -37,6 +37,7 @@ import {QuantityUnitComponent} from './widgets/quantity-unit/quantity-unit.compo
 import {NumberComponent} from './widgets/number/number.component';
 import {TerminologyServerComponent} from './widgets/terminology-server/terminology-server.component';
 import {AnswerValueSetComponent} from './widgets/answer-value-set/answer-value-set.component';
+import {ItemControlComponent} from './widgets/item-control/item-control.component';
 import {DateComponent} from './widgets/date/date.component';
 import {TextAreaComponent} from './widgets/textarea/textarea.component';
 import {DatetimeComponent} from './widgets/datetime/datetime.component';
@@ -73,6 +74,7 @@ export class LformsWidgetRegistry extends DefaultWidgetRegistry {
     this.register('units', UnitsComponent);
     this.register('answer-option', AnswerOptionComponent);
     this.register('answer-option-methods', AnswerOptionMethodsComponent);
+    this.register('item-control', ItemControlComponent);
     this.register('initial', InitialComponent);
     this.register('help-text', HelpTextComponent);
     this.register('string-with-css', StringWithCssComponent);