+++ src/app/sf-form-wrapper/sf-form-wrapper.component.ts
@@ -79,6 +79,8 @@ export class SfFormWrapperComponent implements OnChanges, AfterViewInit {
   ngOnChanges(changes: SimpleChanges) {
     if(changes.model) {
       this.loading = true;
+      // Notify the changes to the form.
+      this.formService.formChanged(changes.model);
       // The new form properties may not be updated at the time of this invocation. Wait for next tick to notify.
       setTimeout(() => {
         this.formService.resetForm();